Add new exceptions to methods

+++ JFL7XU_HFT_2022232.Logic/Logics/OwnerLogic.cs
@@ -1,5 +1,6 @@
 ﻿using JFL7XU_HFT_2022232.Logic.Interfaces;
 using JFL7XU_HFT_2022232.Models;
+using JFL7XU_HFT_2022232.Models.Exceptions;
 using JFL7XU_HFT_2022232.Repository.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -20,15 +21,28 @@ namespace JFL7XU_HFT_2022232.Logic.Logics
 
         public void Create(Owner item)
         {
+            if (repo.Read(item.ID) is not null)
+            {
+                throw new GivenIDAlreadyExistsException();
+            }
             repo.Create(item);
         }
         public void Delete(int id)
         {
+            if (repo.Read(id) is not null)
+            {
+                throw new GivenIDNotFoundException();
+            }
             repo.Delete(id);
         }
         public Owner Read(int id)
         {
-            return repo.Read(id);
+            var owner = repo.Read(id);
+            if (owner is null)
+            {
+                throw new NoOwnerFoundWithGivenIdException();
+            }
+            return owner;
         }
         public IEnumerable<Owner> ReadAll()
         {
@@ -36,6 +50,10 @@ namespace JFL7XU_HFT_2022232.Logic.Logics
         }
         public void Update(Owner item)
         {
+            if (repo.Read(item.ID) is not null)
+            {
+                throw new GivenIDNotFoundException();
+            }
             repo.Update(item);
         }
     }
